requirements.txt
python-telegram-bot[ext]==22.1

google-genai==1.19.0

pydantic-settings==2.3.4

md2tgmd==0.3.10

# python-dotenv is no longer needed
